# Fix `/api/v1/vulnerability/source/{source}/vuln/{vuln}` returning all vulnerable components