@nscuro 2024-09-29
Testing with GHSA-288c-cq4h-88gq which affects 115 projects in my test portfolio:

<img width="750" alt="image" src="https://github.com/user-attachments/assets/b2f50297-6866-4643-a19b-1fd7aedca20a">


Before (**all** requests timing out after 30s while waiting for a DB connection):

```
$ hey -H 'X-Api-Key: odt_***' -n 1000 http://localhost:8080/api/v1/vulnerability/source/GITHUB/vuln/GHSA-288c-cq4h-88gq

Summary:
  Total:	400.0513 secs
  Slowest:	0.0000 secs
  Fastest:	0.0000 secs
  Average:	 NaN secs
  Requests/sec:	2.4997


Response time histogram:


Latency distribution:

Details (average, fastest, slowest):
  DNS+dialup:	 NaN secs, 0.0000 secs, 0.0000 secs
  DNS-lookup:	 NaN secs, 0.0000 secs, 0.0000 secs
  req write:	 NaN secs, 0.0000 secs, 0.0000 secs
  resp wait:	 NaN secs, 0.0000 secs, 0.0000 secs
  resp read:	 NaN secs, 0.0000 secs, 0.0000 secs

Status code distribution:

Error distribution:
  [1000]	Get "http://localhost:8080/api/v1/vulnerability/source/GITHUB/vuln/GHSA-288c-cq4h-88gq": context deadline exceeded (Client.Timeout exceeded while awaiting headers)
```

After:

```
$ hey -H 'X-Api-Key: odt_***' -n 1000 http://localhost:8080/api/v1/vulnerability/source/GITHUB/vuln/GHSA-288c-cq4h-88gq

Summary:
  Total:	1.0845 secs
  Slowest:	0.2268 secs
  Fastest:	0.0061 secs
  Average:	0.0498 secs
  Requests/sec:	922.0578


Response time histogram:
  0.006 [1]	|
  0.028 [245]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.050 [350]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.072 [227]	|■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.094 [102]	|■■■■■■■■■■■■
  0.116 [38]	|■■■■
  0.139 [17]	|■■
  0.161 [12]	|■
  0.183 [3]	|
  0.205 [4]	|
  0.227 [1]	|


Latency distribution:
  10% in 0.0172 secs
  25% in 0.0285 secs
  50% in 0.0437 secs
  75% in 0.0656 secs
  90% in 0.0860 secs
  95% in 0.1083 secs
  99% in 0.1535 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0002 secs, 0.0061 secs, 0.2268 secs
  DNS-lookup:	0.0001 secs, 0.0000 secs, 0.0016 secs
  req write:	0.0000 secs, 0.0000 secs, 0.0014 secs
  resp wait:	0.0492 secs, 0.0055 secs, 0.2267 secs
  resp read:	0.0003 secs, 0.0001 secs, 0.0039 secs

Status code distribution:
  [200]	1000 responses
```